GameMater AI/requirements.txt
langchain-core
langchain-groq
langchain-community
python-dotenv